PyBackEnd/requirements.txt
requests==2.31.0
schedule==1.2.1
python-dotenv==1.0.0
pydantic==2.6.1
aiohttp==3.9.3
asyncio==3.4.3 